+++ payments/src/funcs/documentSigningAndEligibilityList.ts
@@ -4,6 +4,7 @@
 
 import { PaymentsCore } from "../core.js";
 import * as M from "../lib/matchers.js";
+import { compactMap } from "../lib/primitives.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
@@ -44,9 +45,9 @@ export async function documentSigningAndEligibilityList(
     "/payments/collaborator-settings/eligibility-requirement",
   )();
 
-  const headers = new Headers({
+  const headers = new Headers(compactMap({
     Accept: "application/json",
-  });
+  }));
 
   const secConfig = await extractSecurity(client._options.bearerAuth);
   const securityInput = secConfig == null ? {} : { bearerAuth: secConfig };